+++ src/DocumentFragment.php
@@ -4,9 +4,10 @@ declare( strict_types = 1 );
 
 namespace Wikimedia\Dodo;
 
+use Wikimedia\Dodo\Internal\FakeElement;
+use Wikimedia\Dodo\Internal\FilteredElementList;
 use Wikimedia\Dodo\Internal\NamespacePrefixMap;
 use Wikimedia\Dodo\Internal\UnimplementedTrait;
-use Wikimedia\Dodo\Internal\Util;
 
 /**
  * DocumentFragment
@@ -24,16 +25,23 @@ class DocumentFragment extends ContainerNode implements \Wikimedia\IDLeDOM\Docum
 	use \Wikimedia\IDLeDOM\Helper\DocumentFragment;
 
 	/**
-	 * HACK! For compatibilty with W3C test suite, which assumes that an
-	 * access to 'attributes' will return null.
 	 * @param string $name
 	 * @return mixed
 	 */
 	protected function _getMissingProp( string $name ) {
-		if ( $name === 'attributes' ) {
-			return null;
+		switch ( $name ) {
+			case 'attributes':
+				// HACK! For compatibilty with W3C test suite, which
+				// assumes that an access to 'attributes' will return
+				// null.
+				return null;
+			case 'innerHTML':
+				return $this->getInnerHTML(); // nonstandard but handy
+			case 'outerHTML':
+				return $this->getOuterHTML();  // nonstandard but handy
+			default:
+				return parent::_getMissingProp( $name );
 		}
-		return parent::_getMissingProp( $name );
 	}
 
 	/** @inheritDoc */
@@ -80,56 +88,40 @@ class DocumentFragment extends ContainerNode implements \Wikimedia\IDLeDOM\Docum
 		}
 	}
 
-	// Somewhat annoyingly, querySelector/querySelectorAll needs some methods
-	// of Document/Element... but DocumentFragment doesn't have them.
-	// Fake it out by constructing a pseudo-element!
-
 	/** @inheritDoc */
 	public function querySelectorAll( string $selectors ) {
-		$fakeElement = new class( $this ) extends Element {
-			/** @var DocumentFragment $docFrag */
-			private $docFrag;
-
-			/** @param DocumentFragment $docFrag */
-			public function __construct( $docFrag ) {
-				parent::__construct(
-					$docFrag->_nodeDocument,
-					'fake',
-					Util::NAMESPACE_HTML,
-					null
-				);
-				$this->docFrag = $docFrag;
-			}
-
-			// Believe it or not, this is the only method we need to fake
-
-			/** @inheritDoc */
-			public function _nextElement( ?Element $root ): ?Element {
-				return $this->docFrag->getFirstElementChild();
-			}
-		};
-
-		return $fakeElement->querySelectorAll( $selectors );
+		return $this->_fakeElement()->querySelectorAll( $selectors );
 	}
-
-	// Just implement querySelector in terms of querySelectorAll
 
 	/** @inheritDoc */
 	public function querySelector( string $selectors ) {
-		$nodeList = $this->querySelectorAll( $selectors );
-		if ( $nodeList->getLength() > 0 ) {
-			return $nodeList->item( 0 );
-		}
-		return null;
+		return $this->_fakeElement()->querySelector( $selectors );
+	}
+
+	/** @inheritDoc */
+	public function getElementById( string $id ) {
+		$nl = new FilteredElementList( $this->_fakeElement(), static function ( $el ) use ( $id ) {
+			return $el->getAttribute( 'id' ) === $id;
+		} );
+		return $nl->getLength() > 0 ? $nl->item( 0 ) : null;
+	}
+
+	/**
+	 * Create a FakeElement so that we can invoke methods of Element on
+	 * DocumentFragment "as if it were an element".
+	 * @return FakeElement
+	 */
+	private function _fakeElement(): FakeElement {
+		return new FakeElement( $this->_nodeDocument, function () {
+			return $this->getFirstChild();
+		} );
 	}
 
 	// Non-standard, but useful (github issue #73)
 
 	/** @return string the inner HTML of this DocumentFragment */
 	public function getInnerHTML(): string {
-		$result = [];
-		$this->_htmlSerialize( $result );
-		return implode( '', $result );
+		return $this->_fakeElement()->getInnerHTML();
 	}
 
 	/** @return string the outer HTML of this DocumentFragment */